refactor : coop Api 분리
