+++ src/main/java/in/koreatech/koin/domain/coop/controller/CoopApi.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.dining.controller;
+package in.koreatech.koin.domain.coop.controller;
 
 import static in.koreatech.koin.domain.user.model.UserType.COOP;
 
@@ -6,7 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import in.koreatech.koin.domain.dining.dto.DiningImageRequest;
+import in.koreatech.koin.domain.coop.dto.DiningImageRequest;
 import in.koreatech.koin.global.auth.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -16,7 +16,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
 @Tag(name = "(OWNER) Coop Dining : 영양사 식단", description = "영양사 식단 페이지")
-public interface CoopDiningApi {
+public interface CoopApi {
 
     @ApiResponses(
