+++ src/main/java/in/koreatech/koin/domain/coop/controller/CoopController.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.dining.controller;
+package in.koreatech.koin.domain.coop.controller;
 
 import static in.koreatech.koin.domain.user.model.UserType.COOP;
 
@@ -7,16 +7,16 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import in.koreatech.koin.domain.dining.dto.DiningImageRequest;
-import in.koreatech.koin.domain.dining.service.CoopDiningService;
+import in.koreatech.koin.domain.coop.dto.DiningImageRequest;
+import in.koreatech.koin.domain.coop.service.CoopService;
 import in.koreatech.koin.global.auth.Auth;
 import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
-public class CoopDiningController implements CoopDiningApi {
+public class CoopController implements CoopApi {
 
-    private final CoopDiningService coopDiningService;
+    private final CoopService coopDiningService;
 
     @PatchMapping("/coop/dining/image")
     public ResponseEntity<Void> saveDiningImage(
